Added more DB functionality

+++ code/backend/src/main/java/myscores/repositories/GamblerRepository.java
@@ -3,13 +3,16 @@ package myscores.repositories;
 import myscores.database.LocalGraphDatabase;
 import myscores.database.Props;
 import myscores.domain.Gambler;
+import myscores.mappers.GamblerMapper;
 import org.neo4j.graphdb.Node;
 import org.neo4j.graphdb.Transaction;
 import org.neo4j.graphdb.index.Index;
+import org.neo4j.graphdb.index.IndexHits;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.inject.Inject;
+import java.util.ArrayList;
 import java.util.List;
 
 public class GamblerRepository extends Repository<Gambler> {
@@ -19,17 +22,21 @@ public class GamblerRepository extends Repository<Gambler> {
     @Inject
     private LocalGraphDatabase database;
 
+    @Inject
+    private GamblerMapper mapper;
+
     @Override
     public Gambler read(int id) {
         LOGGER.info("Read gambler for id {}", id);
-        Gambler gambler;
+        Gambler gambler = null;
         try (Transaction tx = database.startTransaction()) {
             Index<Node> index = database.getNodeIndex(Props.GAMBLERS_INDEX);
             Node node = index.get(Props.ID, id).getSingle();
-            gambler = new Gambler();
-            gambler.setId((Integer) node.getProperty(Props.ID));
-            gambler.setName((String) node.getProperty(Props.NAME));
-            gambler.setActive((Boolean) node.getProperty(Props.ACTIVE));
+            if (node != null) {
+                gambler = mapper.mapNode(node);
+            } else {
+                LOGGER.warn("Read failed. No gambler found for id {}", id);
+            }
             tx.success();
         }
         return gambler;
@@ -38,7 +45,20 @@ public class GamblerRepository extends Repository<Gambler> {
     @Override
     public List<Gambler> find() {
         LOGGER.info("Find gamblers");
-        return null;
+        List<Gambler> gamblers = new ArrayList<>();
+        try (Transaction tx = database.startTransaction()) {
+            Index<Node> index = database.getNodeIndex(Props.GAMBLERS_INDEX);
+            IndexHits<Node> nodes = index.query(Props.ID, Props.ALL);
+            if (nodes != null) {
+                for (Node node : nodes) {
+                    gamblers.add(mapper.mapNode(node));
+                }
+            } else {
+                LOGGER.warn("Find failed. No gamblers found");
+            }
+            tx.success();
+        }
+        return gamblers;
     }
 
     @Override
@@ -46,12 +66,13 @@ public class GamblerRepository extends Repository<Gambler> {
         LOGGER.info("Create gambler with id {}", gambler.getId());
         try (Transaction tx = database.startTransaction()) {
             Index<Node> index = database.getNodeIndex(Props.GAMBLERS_INDEX);
-            Node node = database.createNode();
-            node.setProperty(Props.ID, gambler.getId());
-            node.setProperty(Props.NAME, gambler.getName());
-            node.setProperty(Props.ACTIVE, gambler.isActive());
-            index.add(node, Props.ID, gambler.getId());
-            index.add(node, Props.NAME, gambler.getName());
+            if (index.get(Props.ID, gambler.getId()).getSingle() == null) {
+                Node node = mapper.mapNode(database.createNode(), gambler);
+                index.add(node, Props.ID, gambler.getId());
+                index.add(node, Props.NAME, gambler.getName());
+            } else {
+                LOGGER.warn("Create failed. Gambler with id {} already exists", gambler.getId());
+            }
             tx.success();
         }
     }
@@ -62,8 +83,12 @@ public class GamblerRepository extends Repository<Gambler> {
         try (Transaction tx = database.startTransaction()) {
             Index<Node> index = database.getNodeIndex(Props.GAMBLERS_INDEX);
             Node node = index.get(Props.ID, gambler.getId()).getSingle();
-            node.setProperty(Props.NAME, gambler.getName());
-            node.setProperty(Props.ACTIVE, gambler.isActive());
+            if (node != null) {
+                node.setProperty(Props.NAME, gambler.getName());
+                node.setProperty(Props.ACTIVE, gambler.isActive());
+            } else {
+                LOGGER.warn("Update failed. No gambler found for id {}", gambler.getId());
+            }
             tx.success();
         }
     }
@@ -74,10 +99,16 @@ public class GamblerRepository extends Repository<Gambler> {
         try (Transaction tx = database.startTransaction()) {
             Index<Node> index = database.getNodeIndex(Props.GAMBLERS_INDEX);
             Node node = index.get(Props.ID, id).getSingle();
-            index.remove(node, Props.ID, node.getProperty(Props.ID));
-            index.remove(node, Props.NAME, node.getProperty(Props.NAME));
-            node.delete();
+            if (node != null) {
+                index.remove(node, Props.ID, node.getProperty(Props.ID));
+                index.remove(node, Props.NAME, node.getProperty(Props.NAME));
+                node.delete();
+            } else {
+                LOGGER.warn("Delete failed. No gambler found for id {}", id);
+            }
             tx.success();
         }
     }
+
+
 }
