+++ code/backend/src/main/java/myscores/repositories/TeamRepository.java
@@ -3,13 +3,16 @@ package myscores.repositories;
 import myscores.database.LocalGraphDatabase;
 import myscores.database.Props;
 import myscores.domain.Team;
+import myscores.mappers.TeamMapper;
 import org.neo4j.graphdb.Node;
 import org.neo4j.graphdb.Transaction;
 import org.neo4j.graphdb.index.Index;
+import org.neo4j.graphdb.index.IndexHits;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.inject.Inject;
+import java.util.ArrayList;
 import java.util.List;
 
 public class TeamRepository extends Repository<Team> {
@@ -19,15 +22,21 @@ public class TeamRepository extends Repository<Team> {
     @Inject
     private LocalGraphDatabase database;
 
+    @Inject
+    private TeamMapper mapper;
+
     @Override
     public Team read(int id) {
         LOGGER.info("Read team for id {}", id);
-        Team team;
+        Team team = null;
         try (Transaction tx = database.startTransaction()) {
             Index<Node> index = database.getNodeIndex(Props.TEAMS_INDEX);
             Node node = index.get(Props.ID, id).getSingle();
-            team = new Team();
-            team.setId((Integer) node.getProperty(Props.ID));
+            if (node != null) {
+                team = mapper.mapNode(node);
+            } else {
+                LOGGER.warn("Read failed. No team found for id {}", id);
+            }
             tx.success();
         }
         return team;
@@ -36,7 +45,20 @@ public class TeamRepository extends Repository<Team> {
     @Override
     public List<Team> find() {
         LOGGER.info("Find teams");
-        return null;
+        List<Team> teams = new ArrayList<>();
+        try (Transaction tx = database.startTransaction()) {
+            Index<Node> index = database.getNodeIndex(Props.TEAMS_INDEX);
+            IndexHits<Node> nodes = index.query(Props.ID, Props.ALL);
+            if (nodes != null) {
+                for (Node node : nodes) {
+                    teams.add(mapper.mapNode(node));
+                }
+            } else {
+                LOGGER.warn("Find failed. No teams found");
+            }
+            tx.success();
+        }
+        return teams;
     }
 
     @Override
@@ -44,9 +66,12 @@ public class TeamRepository extends Repository<Team> {
         LOGGER.info("Create team with id {}", team.getId());
         try (Transaction tx = database.startTransaction()) {
             Index<Node> index = database.getNodeIndex(Props.TEAMS_INDEX);
-            Node node = database.createNode();
-            node.setProperty(Props.ID, team.getId());
-            index.add(node, Props.ID, team.getId());
+            if (index.get(Props.ID, team.getId()).getSingle() == null) {
+                Node node = mapper.mapNode(database.createNode(), team);
+                index.add(node, Props.ID, team.getId());
+            } else {
+                LOGGER.warn("Create failed. Team with id {} already exists", team.getId());
+            }
             tx.success();
         }
     }
@@ -54,6 +79,16 @@ public class TeamRepository extends Repository<Team> {
     @Override
     public void update(Team team) {
         LOGGER.info("Update team with id {}", team.getId());
+        try (Transaction tx = database.startTransaction()) {
+            Index<Node> index = database.getNodeIndex(Props.TEAMS_INDEX);
+            Node node = index.get(Props.ID, team.getId()).getSingle();
+            if (node != null) {
+                node.setProperty(Props.NAME, team.getName());
+            } else {
+                LOGGER.warn("Update failed. No team found for id {}", team.getId());
+            }
+            tx.success();
+        }
     }
 
     @Override
@@ -62,8 +97,12 @@ public class TeamRepository extends Repository<Team> {
         try (Transaction tx = database.startTransaction()) {
             Index<Node> index = database.getNodeIndex(Props.TEAMS_INDEX);
             Node node = index.get(Props.ID, id).getSingle();
-            index.remove(node, Props.ID, node.getProperty(Props.ID));
-            node.delete();
+            if (node != null) {
+                index.remove(node, Props.ID, node.getProperty(Props.ID));
+                node.delete();
+            } else {
+                LOGGER.warn("Delete failed. No team found for id {}", id);
+            }
             tx.success();
         }
     }
